+++ ts/src/profilers/time-profiler-bindings.ts
