+++ ts/test/test-heap-profiler.ts
@@ -16,7 +16,7 @@
 
 import * as sinon from 'sinon';
 import * as heapProfiler from '../src/profilers/heap-profiler';
-import * as v8HeapProfiler from '../src/profilers/heap-profiler-wrapper';
+import * as v8HeapProfiler from '../src/profilers/heap-profiler-bindings';
 import {heapProfileExcludePath, heapProfileIncludePath, heapProfileWithExternal, v8HeapProfile, v8HeapWithPathProfile} from './profiles-for-tests';
 
 const copy = require('deep-copy');
