+++ ts/test/test-time-profiler.ts
@@ -18,7 +18,7 @@ import delay from 'delay';
 import * as sinon from 'sinon';
 import {perftools} from '../../proto/profile';
 import {TimeProfiler} from '../src/profilers/time-profiler';
-import * as v8TimeProfiler from '../src/profilers/time-profiler-wrapper';
+import * as v8TimeProfiler from '../src/profilers/time-profiler-bindings';
 import {timeProfile, v8TimeProfile} from './profiles-for-tests';
 
 const assert = require('assert');
